rating being stored and retrieved in property profile.

+++ RateMyLandlord/RateMyLandlord/Models/Data/Property_Rating.cs
@@ -11,11 +11,12 @@ namespace RateMyLandlord.Models.Data
     public class Property_Rating
     {
         [Key]
+        //[DatabaseGenerated(DatabaseGeneratedOption.None)]
         public int Id { get; set; }
         public int UserId { get; set; }
         public int PropertyId { get; set; }
-        public int pRating { get; set; }
+        public double pRating { get; set; }
         public string Comment { get; set; }
-        public float MonthlyRent { get; set; }
+        public decimal MonthlyRent { get; set; }
     }
 }
